python_src/lessons/pyproject.toml
[tool.poetry]
name = "lessons"
version = "0.1.0"
description = ""
authors = ["Halil Agin <halil.agin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11, <3.12"
pyyaml = "^6.0.2"
pyspark = "3.3.1"
boto3 = "^1.35.88"
jupyter = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
